feat: generate thumbnails at import
