+++ src/components/FileListItem.tsx
@@ -56,6 +56,7 @@ export const FileListItem = memo(FileListItemComponent, (prev, next) => {
   return (
     prev.file.id === next.file.id &&
     prev.file.updatedAt === next.file.updatedAt &&
+    prev.file.objects === next.file.objects &&
     prev.onPressItem === next.onPressItem
   )
 })
